check whether the directory exist

+++ build.py
@@ -43,7 +43,8 @@ project.get('project')['version'] = version_string
 with open('pyproject.toml', 'w', encoding='utf8') as f:
     new_toml_string = toml.dump(project, f)
 
-shutil.rmtree("dist")
+if os.path.isdir("dist"):
+    shutil.rmtree("dist")
 
 os.system(f"git commit -m 'new version {version_string}' pyproject.toml")
 os.system("git push")
